Removed explicit cast since does not seem to support needed binary operations
git-svn-id: http://calder.sdml.cs.kent.edu/svn/srcML/trunk@15114 f70b1346-9fe7-0310-9bf3-95a0de396e6b

+++ src/srcMLapps.py
@@ -10,16 +10,16 @@ LANGUAGE_ASPECTJ = ("AspectJ", 16)
 LANGUAGE_CSHARP = ("C#", 32)
 
 # parsing options
-OPTION_DEBUG           = c_ulonglong(1 << 0)
-OPTION_NESTED          = c_ulonglong(1 << 1)
-OPTION_LITERAL         = c_ulonglong(1 << 2)
-OPTION_CPP_MARKUP_ELSE = c_ulonglong(1 << 22)
+OPTION_DEBUG           = 1 << 0
+OPTION_NESTED          = 1 << 1
+OPTION_LITERAL         = 1 << 2
+OPTION_CPP_MARKUP_ELSE = 1 << 22
 OPTION_XPATH_TOTAL     = OPTION_CPP_MARKUP_ELSE
-OPTION_CPP_MARKUP_IF0  = c_ulonglong(1 << 23)
-OPTION_EXPRESSION      = c_ulonglong(1 << 25)
-OPTION_CPP             = c_ulonglong(1 << 28)
-OPTION_OPERATOR        = c_ulonglong(1 << 16)
-OPTION_MODIFIER        = c_ulonglong(1 << 17)
+OPTION_CPP_MARKUP_IF0  = 1 << 23
+OPTION_EXPRESSION      = 1 << 25
+OPTION_CPP             = 1 << 28
+OPTION_OPERATOR        = 1 << 16
+OPTION_MODIFIER        = 1 << 17
 
 # default uri
 URI_TYPE = c_char_p * 8
